Tesis1_Final/ancash-contract-networks/requirements.txt
import pandas
networkx
matplotlib
jupyter
notebook
numpy
scipy
plotly
pandas_profiling
pyvis
seaborn